moved headless tests under Selenium Examples (#93)
* moved headless tests under Selenium Examples

* deleted headless from pom.xml

* removed headless cross browser test

* changed headless to use sauce bindings
