+++ pom.xml
@@ -22,7 +22,6 @@
         <module>on-boarding-modules/testng</module>
         <module>parallel-testing/cross-browser-and-platform-config</module>
         <module>parallel-testing/single-browser-config</module>
-        <module>headless</module>
         <module>appium-examples</module>
         <module>w3c-examples/w3c-junit5</module>
         <module>w3c-examples/w3c-testng</module>
