+++ selenium-examples/README.md
@@ -2,7 +2,9 @@
 This folder contains Selenium examples
 
 ## Examples
-- [Simple web test, JUnit 5](../selenium-examples/src/test/java/com/saucedemo/JUnit5W3CChromeTest.java)
+
+- [Simple web test, Junit 5](../selenium-examples/src/test/java/com/saucedemo/JUnit5W3CChromeTest.java)
+- [Simple Headless test, TestNG](../selenium-examples/src/test/java/com/saucedemo/SampleHeadlessSauceTest.java)
 
 ## How to run tests
